EDATOS-2928 - No eliminar los atributos cuando se elimina uno de los ficheros de la fuente de datos en un dataset
> Se crea ventana emergente para solicitar al usuario si se deben o no borrar los atributos del dataset al borrar datasources
> Se modifican la fachada y el servicio para pasar el parámetro que indica si se deben o no borrar los atributos
> Se modifican los tests para mantener el comportamiento actual: se eliminan los atributos al borrar datasources

+++ metamac-statistical-resources-core/src/main/java/org/siemac/metamac/statistical/resources/core/dataset/serviceimpl/DatasetServiceImpl.java
@@ -229,10 +229,10 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
     }
 
     @Override
-    public void deleteDatasource(ServiceContext ctx, String urn) throws MetamacException {
+    public void deleteDatasource(ServiceContext ctx, String urn, boolean deleteAttributes) throws MetamacException {
 
         // Validation
-        datasetServiceInvocationValidator.checkDeleteDatasource(ctx, urn);
+        datasetServiceInvocationValidator.checkDeleteDatasource(ctx, urn, deleteAttributes);
 
         // Retrieve
         Datasource datasource = getDatasourceRepository().retrieveByUrn(urn);
@@ -251,7 +251,9 @@ public class DatasetServiceImpl extends DatasetServiceImplBase {
 
         deleteDatasourceData(datasetVersion.getDatasetRepositoryId(), datasource);
 
-        deleteAttributeInstancesLowerThanDatasetLevel(datasetVersion);
+        if (deleteAttributes) {
+            deleteAttributeInstancesLowerThanDatasetLevel(datasetVersion);
+        }
 
         computeDataRelatedMetadata(datasetVersion);
